bill of materials in run mode
