+++ src/components/main-routes/CreateMode.jsx
@@ -6,7 +6,7 @@ import TopMenu from "../secondary/TopMenu.jsx";
 import FlowCanvas from "./flowCanvas.jsx";
 import LowerHalf from "./lowerHalf.jsx";
 import ParamsEditor from "../secondary/ParameterEditor.jsx";
-import { BOMEntry } from "../../js/BOM.js";
+
 import CodeWindow from "../secondary/codeWindow.jsx";
 import {
   BrowserRouter as Router,
@@ -14,7 +14,7 @@ import {
   useNavigate,
 } from "react-router-dom";
 import NewProjectPopUp from "../secondary/NewProjectPopUp.jsx";
-import { button } from "leva";
+
 /**
  * Create mode component appears displays flow canvas, renderer and sidebar when
  * a user has been authorized access to a project.
@@ -33,6 +33,8 @@ function CreateMode(props) {
   const exportPopUp = props.props.exportPopUp;
   const setExportPopUp = props.props.setExportPopUp;
 
+  const compiledBom = props.props.compiledBom;
+
   /** State for grid and axes parameters */
   const [gridParam, setGrid] = useState(true);
   const [axesParam, setAxes] = useState(true);
@@ -45,7 +47,6 @@ function CreateMode(props) {
 
   /** State for top level molecule */
   const [currentMoleculeTop, setTop] = useState(false);
-  const [compiledBom, setCompiledBom] = useState({});
 
   /**
    * Object containing letters and values used for keyboard shortcuts
@@ -95,32 +96,6 @@ function CreateMode(props) {
     //Clearing the interval
     return () => clearInterval(myInterval);
   }, []);
-
-  /** Compile BOM when activeAtom is a molecule and sets new state to trigger menu rerender */
-  /** Should this be in flowcanvas so that it runs after project is loaded? */
-  useEffect(() => {
-    if (GlobalVariables.currentMolecule != undefined && activeAtom) {
-      if (activeAtom.atomType == "Molecule") {
-        compileBom().then((result) => {
-          let bomParams = {};
-          if (result != undefined) {
-            result.map((item) => {
-              bomParams[item.BOMitemName] = {
-                value: item.costUSD + " USD",
-                label: item.BOMitemName + "(x" + item.numberNeeded + ")",
-                disabled: true,
-              };
-            });
-            bomParams["Download List of Materials"] = button(() =>
-              console.log(result)
-            );
-
-            setCompiledBom(bomParams);
-          }
-        });
-      }
-    }
-  }, [activeAtom]);
 
   const handleBodyClick = (e) => {
     if (e.metaKey && e.key == "s") {
@@ -224,39 +199,6 @@ function CreateMode(props) {
     }
   };
 
-  const compileBom = async () => {
-    if (GlobalVariables.currentMolecule.output.value) {
-      let compiled = GlobalVariables.currentMolecule
-        .extractBomTags(GlobalVariables.currentMolecule.output.value)
-        .then((result) => {
-          let bomList = [];
-          let compileBomItems = [];
-          if (result) {
-            result.forEach(function (bomElement) {
-              if (!bomList[bomElement.BOMitemName]) {
-                //If the list of items doesn't already have one of these
-                bomList[bomElement.BOMitemName] = new BOMEntry(); //Create one
-                bomList[bomElement.BOMitemName].numberNeeded = 0; //Set the number needed to zerio initially
-                bomList[bomElement.BOMitemName].BOMitemName =
-                  bomElement.BOMitemName; //With the information from the item
-                bomList[bomElement.BOMitemName].source = bomElement.source;
-                compileBomItems.push(bomList[bomElement.BOMitemName]);
-              }
-              bomList[bomElement.BOMitemName].numberNeeded +=
-                bomElement.numberNeeded;
-              bomList[bomElement.BOMitemName].costUSD += bomElement.costUSD;
-            });
-
-            // Alphabetize by source
-            compileBomItems = compileBomItems.sort((a, b) =>
-              a.source > b.source ? 1 : b.source > a.source ? -1 : 0
-            );
-            return compileBomItems;
-          }
-        });
-      return compiled;
-    }
-  };
   /**
    * Saves project by making a commit to the Github repository.
    */
