+++ src/styles/maslowCreate.css
@@ -843,7 +843,7 @@ Sidebar navigation
 }
 .bomEditorDivRun {
   position: absolute;
-  top: 15%;
+  top: 47%;
   right: 10px;
   width: 280px;
   z-index: 3;
